ks_infrastructure/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ks-infrastructure"
version = "1.0.0"
description = "KS Infrastructure - A unified infrastructure service module"
readme = "README.md"
authors = [{name = "KS Team", email = "ks-team@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "mysql-connector-python>=8.0.0",
    "boto3>=1.26.0",
    "qdrant-client>=1.1.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9",
]

[project.urls]
Homepage = "https://github.com/ks-team/ks-infrastructure"
Repository = "https://github.com/ks-team/ks-infrastructure"